item batches working with custom made units

+++ database/migrations/2023_11_19_043654_create_inventory_items_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('expiration_reminder');
             $table->unsignedBigInteger('category_id');
             $table->unsignedBigInteger('unit_id');
-            $table->integer('total_stock')->nullable()->default(0);
+            $table->decimal('total_stock', 10, 2)->nullable()->default(0);
 
             $table->timestamps();
             $table->unsignedBigInteger('updated_by')->nullable();
